pyproject.toml
[project]
name = "smartsheet-agent"
version = "0.2.0"
description = "AI agent for querying Smartsheet data - jobs, status updates, KPIs, and metrics. Built with Agno framework and supports multiple LLM providers via OpenRouter."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=1.0.0",
    "openai>=1.0.0",
    "smartsheet-python-sdk>=3.2.0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
smartsheet-agent = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "pandas>=2.0.0",
    "jupyter>=1.0.0",
]

[dependency-groups]
dev = [
    "langwatch>=0.7.1",
    "langwatch-scenario>=0.7.14",
    "pytest-asyncio>=1.3.0",
    "pandas>=2.0.0",
]
